refactor: cashouts component
