+++ src/components/cards/wallet/_partials/Cashout.tsx
@@ -7,8 +7,9 @@ import { useSelector } from "@/hooks/useTypedSelector";
 import { openVerificationModal } from "@/store/feature/kyc.slice";
 import { Wallet } from "@/types/wallet";
 import { ReactElement, useCallback, useMemo } from "react";
+import AddressDisplay from "./AddressDisplay";
 
-interface CashoutAddressProps {
+interface CashoutProps {
   wallet: Wallet;
   setShowEditModal: (show: boolean) => void;
   disabled: boolean;
@@ -21,12 +22,11 @@ interface CashoutAddressProps {
  *
  * @returns {ReactElement}
  */
-export default function CashoutAddress({ wallet, setShowEditModal, disabled, setShowPayoutModal, testId = "cashoutAddressId" }: CashoutAddressProps): ReactElement {
+export default function Cashout({ wallet, setShowEditModal, disabled, setShowPayoutModal, testId = "cashoutAddressId" }: CashoutProps): ReactElement {
   const { t } = useTranslation();
   const dispatch = useDispatch();
   const user = useSelector((state) => state.user.data);
-  const isKycVerified = user?.kycStatus === "VERIFIED";
-  const address = useMemo(() => (wallet.address ? wallet.address.match(/.{1,4}/g) : null), [wallet.address]);
+  const isKycVerified = useMemo(() => user?.kycStatus === "VERIFIED", [user?.kycStatus]);
   const cashable = useMemo(() => String(wallet.token).toUpperCase() !== "DAC", [wallet.token]);
 
   const triggerEditAddress = useCallback(() => {
@@ -58,24 +58,7 @@ export default function CashoutAddress({ wallet, setShowEditModal, disabled, set
   return (
     <div className="px-7 pt-6 flex-1 pb-24 lg:pb-24" data-testid={testId}>
       {cashable ? (
-        <div className="text-sm text-gray-700">
-          {address ? (
-            <p className="leading-5 text-sm flex gap-x-2 gap-y-1 flex-wrap font-mono font-normal">
-              {address.map((part, index) => (
-                <span key={`address-${index}`} className="mr-2">
-                  {part}
-                </span>
-              ))}
-            </p>
-          ) : (
-            <p>{wallet.description}</p>
-          )}
-          <div className="text-gray-700 text-sm mt-3">
-            <span className="cursor-pointer hover:underline" onClick={triggerEditAddress}>
-              {address ? t("profile.wallets.address-change") : t("profile.wallets.address-set")}
-            </span>
-          </div>
-        </div>
+        <AddressDisplay walletAddress={wallet.address} description={wallet.description} triggerEditAddress={triggerEditAddress} />
       ) : (
         <div className="prose">
           <p
